moved advancedLogic from controllers to another file; changed error handling in dvancedLogic; added tests for checkFigureColor, getSquaresToBePassed & checkSquaresToBePassed
