+++ validation/advancedLogic_test.go
@@ -1,0 +1,97 @@
+package validation
+
+import (
+	"testing"
+
+	"github.com/sadmadrus/chessBox/internal/board"
+)
+
+func TestCheckFigureColor(t *testing.T) {
+	tests := []struct {
+		name  string
+		brd   board.Board
+		piece board.Piece
+		isOk  bool
+	}{
+		{"R classical", *board.Classical(), board.WhiteRook, true},
+		{"P classical", *board.Classical(), board.WhitePawn, true},
+		{"k classical", *board.Classical(), board.BlackKing, false},
+		{"b classical", *board.Classical(), board.BlackBishop, false},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			res := checkFigureColor(tc.brd, tc.piece)
+			if res != tc.isOk {
+				t.Fatalf("want %v, got %v", tc.isOk, res)
+			}
+		})
+	}
+}
+
+func TestGetSquaresToBePassed(t *testing.T) {
+	tests := []struct {
+		name   string
+		piece  board.Piece
+		from   square
+		to     square
+		result []square
+	}{
+		{"P e2-e4", board.WhitePawn, newSquare(12), newSquare(28), []square{newSquare(20)}},
+		{"p a7-a6", board.BlackPawn, newSquare(48), newSquare(40), []square{}},
+		{"N g1-h3", board.WhiteKnight, newSquare(6), newSquare(23), []square{}},
+		{"n b8-d7", board.BlackKnight, newSquare(57), newSquare(51), []square{}},
+		{"B f1-a6", board.WhiteBishop, newSquare(5), newSquare(40),
+			[]square{newSquare(12), newSquare(19), newSquare(26), newSquare(33)}},
+		{"b c8-a6", board.BlackBishop, newSquare(58), newSquare(40), []square{newSquare(49)}},
+		{"R h1-e1", board.WhiteRook, newSquare(7), newSquare(4),
+			[]square{newSquare(6), newSquare(5)}},
+		{"r a8-a7", board.BlackRook, newSquare(56), newSquare(48), []square{}},
+		{"Q d1-d4", board.WhiteQueen, newSquare(3), newSquare(27),
+			[]square{newSquare(11), newSquare(19)}},
+		{"q h8-f6", board.BlackQueen, newSquare(63), newSquare(45), []square{newSquare(54)}},
+		{"K e5-e6", board.WhiteKing, newSquare(36), newSquare(44), []square{}},
+		{"k e8-g8 O-O", board.BlackKing, newSquare(60), newSquare(62), []square{newSquare(61)}},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			res := getSquaresToBePassed(tc.piece, tc.from, tc.to)
+			for i, _ := range res {
+				if res[i] != tc.result[i] {
+					t.Fatalf("want %v, got %v", tc.result, res)
+				}
+			}
+		})
+	}
+}
+
+func TestCheckSquaresToBePassed(t *testing.T) {
+	tests := []struct {
+		name    string
+		brd     board.Board
+		squares []square
+		result  bool
+	}{
+		{"a3-a6 classical", *board.Classical(),
+			[]square{newSquare(16), newSquare(24), newSquare(32), newSquare(40)}, true},
+		{"a6-a7 classical", *board.Classical(), []square{newSquare(40), newSquare(48)}, false},
+		{"h8 classical", *board.Classical(), []square{newSquare(63)}, false},
+		{"h3 classical", *board.Classical(), []square{newSquare(23)}, true},
+		{"a1-b2 classical", *board.Classical(), []square{newSquare(0), newSquare(9)}, false},
+		{"h4-g5 classical", *board.Classical(), []square{newSquare(31), newSquare(38)}, true},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			res, err := checkSquaresToBePassed(tc.brd, tc.squares)
+			if err != nil {
+				t.Fatalf("want nil, got error: %s", err)
+			}
+
+			if res != tc.result {
+				t.Fatalf("want %v, got %v", tc.result, res)
+			}
+		})
+	}
+}
